Multiple UI changes
Changed multiple UI components including most buttons to use MUI components. Along with this I restructured the files to help with the expanding component list. This way adding a page or something to a page will be more organized. Finally I added a route in the backend. I also deleted the Logout component because the functionality is stored in the header.