Profile and Styling Updates
Added a simple Profile page that is a work in progress. Currently has your Full name and Email along with all the groups you are the leader of. In the next commit I plan to change that to all groups you are a member of. Along with profile changes, I also added a sticky position for the create a group modal button to allow for easier access when using the site.